downlit-package {downlit}	R Documentation

downlit: Syntax Highlighting and Automatic Linking

Description

Syntax highlighting of R code, specifically designed for the needs of 'RMarkdown' packages like 'pkgdown', 'hugodown', and 'bookdown'. It includes linking of function calls to their documentation on the web, and automatic translation of ANSI escapes in output to the equivalent HTML.

Author(s)

Maintainer: Hadley Wickham hadley@posit.co

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://downlit.r-lib.org/

https://github.com/r-lib/downlit

Report bugs at https://github.com/r-lib/downlit/issues

[Package downlit version 0.4.3 Index]